requirements.txt
Flask==3.0.0
flask-restx==1.3.0
python-dotenv==1.0.0
flask-limiter==3.5.0
werkzeug==3.0.1
pytest==7.4.3
gunicorn==21.2.0
python-magic==0.4.27; platform_system != "Darwin"
python-magic-bin==0.4.14; platform_system == "Darwin"
requests==2.31.0
